move loader page to common components
