+++ client/src/components/log/log.tsx
@@ -1,7 +1,7 @@
 import { useGW2Log } from '../../lib/apis/gw2-api';
 import { useFilterString } from '../../lib/utils/use-filter-string';
 import { ErrorMessage } from '../common/error-message';
-import LoaderPage from '../loader-page';
+import LoaderPage from '../common/loader-page';
 import './log.scss';
 import LogEntry from './log-entry';
